requirements.txt
numpy==1.23.4
Pillow==9.3.0
